move create vtop release to release step instead of pre-release
