+++ go/interactive/release/vtop_create_release_pr.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package pre_release
+package release
 
 import (
 	"context"
